Rename api.rb to url_helper_methods
