+++ app/controllers/application_controller.rb
@@ -1,4 +1,4 @@
-require_relative 'api'
+require_relative 'url_helper_methods'
 
 class ApplicationController < ActionController::API
 end
